requirements.txt
flask
flask-cors
requests
python-dotenv
pillow
numpy
opencv-python-headless